+++ src/main/java/mServer/crawler/sender/base/UrlUtils.java
@@ -5,7 +5,6 @@ import java.io.IOException;
 import java.util.HashMap;
 import java.util.Map;
 import java.util.Optional;
-import mServer.crawler.sender.phoenix.UrlParseException;
 import okhttp3.Request;
 import okhttp3.Response;
 import org.jetbrains.annotations.NotNull;
